+++ packages/stream_video_flutter/lib/src/call_controls/controls/stream_closed_captions_button.dart
@@ -2,11 +2,10 @@ import 'package:flutter/material.dart';
 
 import '../../../stream_video_flutter.dart';
 
-/// A widget that represents a call control option to toggle if closed captions
-/// is on or off.
-class ToggleClosedCaptionsOption extends StatelessWidget {
-  /// Creates a new instance of [ToggleClosedCaptionsOption].
-  const ToggleClosedCaptionsOption({
+/// A call control that turns closed captions on and off.
+class StreamClosedCaptionsButton extends StatelessWidget {
+  /// Creates a new instance of [StreamClosedCaptionsButton].
+  const StreamClosedCaptionsButton({
     super.key,
     required this.call,
     this.enabledClosedCaptionIcon,
@@ -53,3 +52,10 @@ class ToggleClosedCaptionsOption extends StatelessWidget {
     );
   }
 }
+
+/// ToggleClosedCaptionsOption is [StreamClosedCaptionsButton] now.
+@Deprecated(
+  'ToggleClosedCaptionsOption is StreamClosedCaptionsButton now, matching the rest of the '
+  'call controls. Will be removed in the next major version.',
+)
+typedef ToggleClosedCaptionsOption = StreamClosedCaptionsButton;