+++ packages/stream_video_flutter/lib/src/l10n/localizations/stream_video_flutter_localizations_nl.dart
@@ -68,10 +68,7 @@ class StreamVideoFlutterLocalizationsNl
   String get lobbyCameraSection => 'Camera';
 
   @override
-  String get lobbySystemDefaultDevice => 'Systeemstandaard';
-
-  @override
-  String get lobbyDefaultDeviceHint => 'Standaard';
+  String get defaultDevice => 'Standaard';
 
   @override
   String get lobbySelectAudioDevices => 'Audioapparaten selecteren';